enhance ui
