+++ src/components/user/journal/journal_list_layout.tsx
@@ -30,9 +30,10 @@ const JournalListItem = ({
     </div>
   );
 
+  // Info: (20260320 - Julian) 尚未開始
   if (journal.analysisStatus === AIAnalysisStatus.PENDING) {
     return (
-      <tr className="border-b last:border-0 border-slate-300 odd:bg-slate-50 even:bg-white text-slate-400">
+      <tr className="border-b border-slate-300 text-slate-400 last:border-0 odd:bg-slate-50 even:bg-white">
         <td className="w-16 px-3 py-2 align-middle sm:w-32 sm:px-6">
           <div className="flex h-12 w-12 items-center justify-center overflow-hidden rounded-lg border border-dashed border-gray-300 bg-gray-50 p-1 sm:h-20 sm:w-20">
             <Loader2 className="h-4 w-4 animate-spin text-orange-400 sm:h-6 sm:w-6" />
@@ -45,7 +46,7 @@ const JournalListItem = ({
         <td className="px-3 py-2 align-middle text-xs sm:px-6 sm:text-sm">
           <span className="flex items-center gap-2 italic">
             <Loader2 className="h-4 w-4 animate-spin text-orange-400" />
-           AI Analyzing...
+            AI Analyzing...
           </span>
         </td>
         <td className="w-12 px-3 py-2 text-right sm:px-6">
@@ -59,6 +60,11 @@ const JournalListItem = ({
         </td>
       </tr>
     );
+  }
+
+  // Info: (20260320 - Julian) 處理中
+  if (journal.analysisStatus === AIAnalysisStatus.PROCESSING) {
+    return <tr></tr>;
   }
 
   return (
