+++ src/components/user/voucher/account_book_selector.tsx
@@ -1,4 +1,4 @@
-'use client';
+"use client";
 
 import { Fragment, useState, useEffect, useMemo } from "react";
 import {
@@ -58,7 +58,9 @@ export default function AccountBookSelector({
         // eslint-disable-next-line react-hooks/set-state-in-effect
         setIsLoading(true);
         setSelectedType(null);
-        request<IApiResponse<IAccountBook>>(`/api/v1/user/account_book/${accountBookId}`)
+        request<IApiResponse<IAccountBook>>(
+          `/api/v1/user/account_book/${accountBookId}`,
+        )
           .then((res) => {
             if (res.payload) {
               setAccountBook(res.payload);
@@ -97,7 +99,7 @@ export default function AccountBookSelector({
       list = list.filter(
         (acc) =>
           acc.code.toLowerCase().includes(lowerKeyword) ||
-          acc.name.toLowerCase().includes(lowerKeyword)
+          acc.name.toLowerCase().includes(lowerKeyword),
       );
     }
     return list;
@@ -129,14 +131,14 @@ export default function AccountBookSelector({
               leaveFrom="opacity-100 translate-y-0 sm:scale-100"
               leaveTo="opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95"
             >
-              <DialogPanel className="relative transform flex flex-col h-[70vh] bg-white text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-3xl rounded-2xl overflow-hidden">
+              <DialogPanel className="relative flex h-[70vh] transform flex-col overflow-hidden rounded-2xl bg-white text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-3xl">
                 {/* Info: (20260317 - Julian) Header */}
                 <div className="flex items-center justify-between border-b border-slate-200 px-6 py-4">
                   <DialogTitle
                     as="h3"
                     className="text-lg font-bold text-slate-800"
                   >
-                  {t("voucher.account_book_selector.title")}
+                    {t("voucher.account_book_selector.title")}
                   </DialogTitle>
                   <button
                     type="button"
@@ -147,7 +149,7 @@ export default function AccountBookSelector({
                     <X size={18} className="stroke-[2.5]" />
                   </button>
                 </div>
-                
+
                 {/* Info: (20260317 - Julian) Body Content */}
                 {isLoading ? (
                   <div className="flex flex-1 items-center justify-center">
@@ -156,10 +158,10 @@ export default function AccountBookSelector({
                 ) : (
                   <div className="flex flex-1 overflow-hidden">
                     {/* Info: (20260317 - Julian) Left Side: Types */}
-                    <div className="w-1/3 border-r border-slate-100 flex flex-col overflow-y-auto p-4 bg-slate-50/50">
+                    <div className="flex w-1/3 flex-col overflow-y-auto border-r border-slate-100 bg-slate-50/50 p-4">
                       <button
                         onClick={() => setSelectedType(null)}
-                        className={`w-full rounded-xl p-3 text-left transition-colors text-sm font-bold mb-1 ${
+                        className={`mb-1 w-full rounded-xl p-3 text-left text-sm font-bold transition-colors ${
                           selectedType === null
                             ? "bg-orange-100 text-orange-600"
                             : "text-slate-600 hover:bg-slate-100"
@@ -171,7 +173,7 @@ export default function AccountBookSelector({
                         <button
                           key={type}
                           onClick={() => setSelectedType(type)}
-                          className={`w-full rounded-xl p-3 text-left transition-colors text-sm font-bold mb-1 ${
+                          className={`mb-1 w-full rounded-xl p-3 text-left text-sm font-bold transition-colors ${
                             selectedType === type
                               ? "bg-orange-100 text-orange-600"
                               : "text-slate-600 hover:bg-slate-100"
@@ -183,21 +185,23 @@ export default function AccountBookSelector({
                     </div>
 
                     {/* Info: (20260317 - Julian) Right Side: Search & List */}
-                    <div className="w-2/3 flex flex-col">
-                      <div className="p-4 border-b border-slate-100">
+                    <div className="flex w-2/3 flex-col">
+                      <div className="border-b border-slate-100 p-4">
                         <div className="relative">
-                          <Search className="absolute left-3 top-1/2 h-4 w-4 -translate-y-1/2 text-slate-400" />
+                          <Search className="absolute top-1/2 left-3 h-4 w-4 -translate-y-1/2 text-slate-400" />
                           <input
                             type="text"
                             aria-label="Search"
                             value={keyword}
                             onChange={(e) => setKeyword(e.target.value)}
-                            placeholder={t("voucher.account_book_selector.search_placeholder")}
-                            className="w-full rounded-xl border border-slate-200 bg-slate-50 py-2.5 pl-9 pr-4 text-sm font-semibold text-slate-800 placeholder:text-slate-400 focus:border-orange-500 focus:bg-white focus:outline-none focus:ring-1 focus:ring-orange-500"
+                            placeholder={t(
+                              "voucher.account_book_selector.search_placeholder",
+                            )}
+                            className="w-full rounded-xl border border-slate-200 bg-slate-50 py-2.5 pr-4 pl-9 text-sm font-semibold text-slate-800 placeholder:text-slate-400 focus:border-orange-500 focus:bg-white focus:ring-1 focus:ring-orange-500 focus:outline-none"
                           />
                         </div>
                       </div>
-                      
+
                       <div className="flex-1 overflow-y-auto p-4">
                         {filteredAccounts.length > 0 ? (
                           <div className="flex flex-col gap-1">
@@ -208,23 +212,29 @@ export default function AccountBookSelector({
                                   onSelect(acc);
                                   onClose();
                                 }}
-                                className="w-full rounded-xl p-3 text-left transition-colors hover:bg-slate-50 disabled:opacity-50 flex items-center gap-3 group border border-transparent hover:border-slate-200"
+                                className="group flex w-full items-center gap-3 rounded-xl border border-transparent p-3 text-left transition-colors hover:border-slate-200 hover:bg-slate-50 disabled:opacity-50"
                               >
-                                <div className="flex h-10 w-10 shrink-0 items-center justify-center rounded-lg bg-slate-100 text-xs font-bold text-slate-500 group-hover:bg-orange-100 group-hover:text-orange-600 transition-colors">
+                                <div className="flex h-10 w-10 shrink-0 items-center justify-center rounded-lg bg-slate-100 text-xs font-bold text-slate-500 transition-colors group-hover:bg-orange-100 group-hover:text-orange-600">
                                   {acc.code}
                                 </div>
                                 <div className="flex flex-col">
-                                  <span className="text-sm font-bold text-slate-700">{acc.name}</span>
-                                  <span className="text-xs font-medium text-slate-400 mt-0.5">{acc.type}</span>
+                                  <span className="text-sm font-bold text-slate-700">
+                                    {acc.name}
+                                  </span>
+                                  <span className="mt-0.5 text-xs font-medium text-slate-400">
+                                    {acc.type}
+                                  </span>
                                 </div>
                               </button>
                             ))}
                           </div>
                         ) : (
                           <div className="flex h-32 flex-col items-center justify-center gap-2 text-slate-400">
-                          <SearchX className="h-8 w-8 text-slate-300" />
-                          <p className="text-sm font-semibold">{t("voucher.account_book_selector.no_results")}</p>
-                        </div>
+                            <SearchX className="h-8 w-8 text-slate-300" />
+                            <p className="text-sm font-semibold">
+                              {t("voucher.account_book_selector.no_results")}
+                            </p>
+                          </div>
                         )}
                       </div>
                     </div>
