+++ src/components/user/voucher/voucher_table_section.tsx
@@ -3,281 +3,18 @@
 import { useState, useEffect, useCallback } from "react";
 import { useParams } from "next/navigation";
 import {
-  ArrowUpRight,
-  ArrowDownLeft,
-  ArrowRightLeft,
   ChevronUp,
   ChevronDown,
   Search,
   CheckCircle2,
-  FileQuestion,
   Filter,
-  Loader2,
 } from "lucide-react";
 import { useTranslation } from "@/i18n/i18n_context";
-import { timestampToString, numberWithCommas } from "@/lib/utils/common";
 import { request } from "@/lib/utils/request";
 import { IApiResponse } from "@/lib/utils/response";
-import { FilePreview } from "@/components/common/file_preview";
+import { VoucherRow } from "@/components/user/voucher/voucher_row";
 import VoucherDetailModal from "@/components/user/voucher/voucher_detail_modal";
 import { IVoucher, TradingType } from "@/interfaces/voucher";
-import { AIAnalysisStatus } from "@/interfaces/ai_analysis_status";
-
-const VoucherRow = ({
-  voucher,
-  onClick,
-}: {
-  voucher: IVoucher;
-  onClick: () => void;
-}) => {
-  const { t } = useTranslation();
-  const lineItems = voucher.lineItems.lines;
-
-  const getMockConfidence = (id: string) => {
-    let sum = 0;
-    for (let i = 0; i < id.length; i++) {
-      sum += id.charCodeAt(i);
-    }
-    return (sum % 15) + 85;
-  };
-  const mockConfidence = getMockConfidence(voucher.id);
-  const mockStatus =
-    parseInt(voucher.id.slice(-1), 16) % 2 === 0 ? "verified" : "manual";
-
-  const renderIcon = (type: TradingType) => {
-    switch (type) {
-      case TradingType.INCOME:
-        return <ArrowDownLeft size={14} className="stroke-[2.5]" />;
-      case TradingType.OUTCOME:
-        return <ArrowUpRight size={14} className="stroke-[2.5]" />;
-      case TradingType.TRANSFER:
-        return <ArrowRightLeft size={14} className="stroke-[2.5]" />;
-      default:
-        return null;
-    }
-  };
-
-  const getTypeClasses = (style: TradingType) => {
-    switch (style) {
-      case TradingType.OUTCOME:
-        return "border-red-200 bg-red-50 text-red-500";
-      case TradingType.INCOME:
-        return "border-emerald-200 bg-emerald-50 text-emerald-500";
-      case TradingType.TRANSFER:
-        return "border-slate-200 bg-slate-50 text-slate-700";
-      default:
-        return "border-slate-200 bg-slate-100 text-slate-600";
-    }
-  };
-
-  const getTypeName = (style: TradingType) => {
-    switch (style) {
-      case TradingType.OUTCOME:
-        return t("voucher.main_view.table.types.outcome") || "支出傳票";
-      case TradingType.INCOME:
-        return t("voucher.main_view.table.types.income") || "收入傳票";
-      case TradingType.TRANSFER:
-        return t("voucher.main_view.table.types.transfer") || "轉帳傳票";
-      default:
-        return t("voucher.main_view.table.types.unknown") || "未知傳票";
-    }
-  };
-
-  if (voucher.analysisStatus === AIAnalysisStatus.PENDING) {
-    return (
-      <tr className="border-b border-slate-300 bg-slate-50 text-sm transition-colors last:border-0 opacity-80">
-        {/* Info: (20260318) File Preview loading */}
-        <td className="p-2 text-center lg:px-6 lg:py-4">
-          <div className="mx-auto flex h-14 w-14 items-center justify-center overflow-hidden rounded-lg border border-dashed border-slate-300 bg-white p-1 shadow-sm sm:h-16 sm:w-16">
-            <Loader2 className="h-6 w-6 animate-spin text-orange-400" />
-          </div>
-        </td>
-        {/* Info: (20260318) Trading Date (still showing the created date conceptually) */}
-        <td className="p-2 text-center align-middle font-bold whitespace-nowrap text-slate-400 lg:px-6 lg:py-4">
-          <p className="text-xs lg:text-sm">
-            {timestampToString(voucher.tradingDate).dateWithDash}
-          </p>
-        </td>
-        {/* Info: (20260318) Colspan the rest of the parsing info to show a generic loading center */}
-        <td colSpan={5} className="p-2 align-middle text-center lg:px-6 lg:py-4">
-          <span className="flex items-center justify-center gap-2 text-sm font-medium italic text-orange-500">
-            <Loader2 className="h-4 w-4 animate-spin text-orange-500" />
-            AI Analyzing...
-          </span>
-        </td>
-        {/* Info: (20260318) Status action (Disabled) */}
-        <td aria-label="Status" className="p-2 text-center align-middle lg:px-6 lg:py-4">
-          <div className="flex flex-col items-center justify-center gap-1.5 ">
-            <button
-              type="button"
-              disabled
-              className="inline-flex cursor-not-allowed items-center justify-center rounded-xl bg-slate-200 px-4 py-1.5 text-sm font-bold whitespace-nowrap text-slate-400 shadow-sm"
-            >
-              Pending
-            </button>
-          </div>
-        </td>
-      </tr>
-    );
-  }
-
-  return (
-    <tr
-      key={voucher.id}
-      className={`border-b border-slate-300 bg-white text-sm transition-colors last:border-0 ${voucher.isDeleted ? "opacity-50" : ""}`}
-    >
-      {/* Info: (20260316 - Julian) File */}
-      <td className="p-2 text-center lg:px-6 lg:py-4">
-        <div className="mx-auto flex h-14 w-14 items-center justify-center overflow-hidden rounded-lg border border-slate-200 bg-white p-1 shadow-sm sm:h-16 sm:w-16">
-          {voucher.file ? (
-            <FilePreview
-              file={{ filename: voucher.file.fileName || "Unknown" }}
-              fileId={voucher.file.hash}
-              className="h-full w-full object-cover"
-            />
-          ) : (
-            <div className="flex h-full w-full flex-col justify-center items-center rounded-lg bg-slate-100 p-1">
-              <FileQuestion className="mb-1 h-5 w-5 text-slate-300" />
-              <span className="text-[10px] font-bold text-slate-400 leading-none">
-                {t("voucher.main_view.table.no_file")}
-              </span>
-            </div>
-          )}
-        </div>
-      </td>
-      {/* Info: (20260316 - Julian) Trading Date */}
-      <td className="p-2 text-center align-middle font-bold whitespace-nowrap text-slate-800 lg:px-6 lg:py-4">
-        <p className="text-xs lg:text-sm">
-          {timestampToString(voucher.tradingDate).dateWithDash}
-        </p>
-        {voucher.isDeleted && (
-          <div className="mt-2 text-center">
-            <span className="inline-block rounded-full bg-orange-200 px-2 py-0.5 text-[10px] font-bold text-orange-500">
-              {t("voucher.main_view.table.status.deleted")}
-            </span>
-          </div>
-        )}
-      </td>
-      {/* Info: (20260316 - Julian) Type */}
-      <td
-        aria-label="Type"
-        className="p-2 text-center align-middle lg:px-6 lg:py-4"
-      >
-        <div className="flex flex-col items-center justify-center gap-2">
-          <div
-            className={`flex items-center gap-1.5 rounded-full border px-3 py-1 text-xs font-bold whitespace-nowrap ${getTypeClasses(voucher.tradingType)}`}
-          >
-            {renderIcon(voucher.tradingType)}
-            <span>{getTypeName(voucher.tradingType)}</span>
-          </div>
-          <span className="text-xs font-black tracking-wider text-slate-800">
-            {voucher.id}
-          </span>
-        </div>
-      </td>
-      {/* Info: (20260316 - Julian) Accounting */}
-      <td
-        aria-label="Accounting"
-        className="py-2 pl-2 align-middle lg:py-4 lg:pl-6"
-      >
-        <div className="flex flex-col whitespace-nowrap">
-          {lineItems.map((line) => (
-            <div
-              key={line.id}
-              className="flex h-[30px] items-center gap-2 border-dashed border-slate-300 not-last:border-b"
-            >
-              <span className="w-[55px] rounded bg-slate-200 px-1.5 py-0.5 text-center text-xs font-semibold text-slate-700">
-                {line.accounting?.code}
-              </span>
-              {/* Info: (20260316 - Julian) 借方靠左，貸方靠右 */}
-              <span
-                className={`${
-                  line.isDebit
-                    ? "font-bold text-slate-800"
-                    : "ml-4 font-medium text-slate-700"
-                } text-xs lg:text-sm truncate lg:max-w-[250px]`}
-              >
-                {line.accounting?.name}
-              </span>
-            </div>
-          ))}
-        </div>
-      </td>
-      {/* Info: (20260316 - Julian) Debit */}
-      <td
-        aria-label="Debit"
-        className="py-2 text-right align-middle font-semibold text-slate-700 lg:py-4"
-      >
-        <div className="flex flex-col text-xs lg:text-sm">
-          {lineItems.map((line) => (
-            <div
-              key={line.id}
-              className="flex h-[30px] items-center justify-end border-dashed border-slate-300 not-last:border-b"
-            >
-              <span>{line.isDebit ? numberWithCommas(line.amount) : "−"}</span>
-            </div>
-          ))}
-        </div>
-      </td>
-      {/* Info: (20260316 - Julian) Credit */}
-      <td
-        aria-label="Credit"
-        className="py-2 pr-2 text-right align-middle font-semibold lg:py-4 lg:pr-6"
-      >
-        <div className="flex flex-col text-xs lg:text-sm">
-          {lineItems.map((line) => (
-            <div
-              key={line.id}
-              className="flex h-[30px] items-center justify-end border-dashed border-slate-300 not-last:border-b"
-            >
-              <span>{!line.isDebit ? numberWithCommas(line.amount) : "−"}</span>
-            </div>
-          ))}
-        </div>
-      </td>
-      {/* Info: (20260316 - Julian) Confidence */}
-      <td
-        aria-label="Confidence"
-        className="p-2 text-center align-middle lg:px-6 lg:py-4"
-      >
-        <div className="flex flex-col-reverse items-center justify-center gap-x-3 gap-y-1 lg:flex-row">
-          <div className="h-1.5 w-16 shrink-0 overflow-hidden rounded-full bg-slate-200">
-            <div
-              className={`h-full rounded-full ${mockConfidence >= 90 ? "bg-emerald-400" : "bg-orange-500"}`}
-              style={{ width: `${mockConfidence}%` }}
-            ></div>
-          </div>
-          <span className="text-sm font-black whitespace-nowrap text-slate-700">
-            {mockConfidence}%
-          </span>
-        </div>
-      </td>
-      {/* Info: (20260316 - Julian) Status */}
-      <td
-        aria-label="Status"
-        className="p-2 text-center align-middle lg:px-6 lg:py-4"
-      >
-        {mockStatus === "verified" ? (
-          <div className="flex flex-col items-center justify-center gap-1 text-emerald-500">
-            <CheckCircle2 className="h-5 w-5" />
-            <span className="text-xs font-bold whitespace-nowrap">{t("voucher.main_view.table.status.verified")}</span>
-          </div>
-        ) : (
-          <div className="flex flex-col items-center justify-center gap-1.5 text-orange-500">
-            <button
-              type="button"
-              onClick={onClick}
-              disabled={voucher.isDeleted}
-              className="inline-flex items-center justify-center rounded-xl bg-orange-500 px-4 py-1.5 text-sm font-bold whitespace-nowrap text-white shadow-sm enabled:hover:bg-orange-600 disabled:bg-slate-300"
-            >
-              {t("voucher.main_view.table.status.manual")}
-            </button>
-          </div>
-        )}
-      </td>
-    </tr>
-  );
-};
 
 enum VoucherSorting {
   DATE_DESC = "date_desc",
@@ -596,7 +333,9 @@ export default function VoucherTableSection() {
                     <th className="p-2 text-center text-xs font-black tracking-wider whitespace-nowrap lg:px-6 lg:py-4">
                       <button
                         type="button"
-                        aria-label={t("voucher.main_view.table.headers.voucher_date")}
+                        aria-label={t(
+                          "voucher.main_view.table.headers.voucher_date",
+                        )}
                         onClick={clickDateSort}
                         className="group mx-auto flex w-full items-center justify-center gap-1"
                       >
@@ -628,7 +367,9 @@ export default function VoucherTableSection() {
                     </th>
                     <th className="p-2 text-left text-xs font-black tracking-wider whitespace-nowrap text-slate-500 uppercase lg:px-6 lg:py-4">
                       <div className="flex w-[250px] items-center">
-                        {t("voucher.main_view.table.headers.accounting_entries")}
+                        {t(
+                          "voucher.main_view.table.headers.accounting_entries",
+                        )}
                       </div>
                     </th>
                     <th className="p-2 text-right text-xs font-black tracking-wider whitespace-nowrap lg:px-6 lg:py-4">
