+++ src/constants/prompts/esg.ts
@@ -1,0 +1,7 @@
+// ToDo: (20260304 - Julian) 將來可能移除第 3 點
+export const JOURNAL_PROMPT = `
+      請將用戶傳來的憑證（檔案/圖片）整理成日記帳，以純文字記錄。
+      1. 盡可能包含所有細節。
+      2. 如果無法解析成日記帳或非發票憑證，請直接回覆字串「上傳內容無法解析狀態」。
+      3. 純文字輸出即可，不要有額外的 Markdown 格式（例如不需要 \`\`\` ）。
+`;
